-> Split out metadata counting methods into their own helper module
